+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+[4.1.0](https://github.com/conductorphp/conductor-azure-blob-filesystem-support/compare/4.0.0...4.1.0) (2026-08-10)
+
+### Features
+* PHP 8.4.1+ (CTAP-1224) ([0d2a049](https://github.com/conductorphp/conductor-azure-blob-filesystem-support/commit/0d2a049891c99f8693132b0447ed2d1325f5da34))
+* laminas-servicemanager v4 (CTAP-1221) ([e598af1](https://github.com/conductorphp/conductor-azure-blob-filesystem-support/commit/e598af1640e83ccaa7e76c40a9906e854bf12998))
+
+<!--- CHANGELOG SPLIT MARKER -->
+
+[3.0.0](https://github.com/conductorphp/conductor-azure-blob-filesystem-support/compare/2.0.0...3.0.0) (2026-06-26)
+
+### Bug Fixes
+* to flysystem 3 (azure-oss) and conductor 2.0 (CTAP-776) ([493d960](https://github.com/conductorphp/conductor-azure-blob-filesystem-support/commit/493d9604be9c5345d2a490b3452922e6a3be5bc4))
+
+<!--- CHANGELOG SPLIT MARKER -->
+
 [1.0.1](https://github.com/conductorphp/conductor-azure-blob-filesystem-support/compare/1.0.0...1.0.1) (2026-06-25)
 
 ### Bug Fixes